requirements.txt
fastapi
mysql-connector-python
uvicorn
flask
requests
numpy
pandas